overhaul with new setup
