--- app/auth/views.py
@@ -1,43 +1,0 @@
-from flask import render_template, redirect, request, url_for, flash
-from flask_login import login_user, logout_user, login_required
-from . import auth
-from .. import db
-from .. models import User
-from .forms import LoginForm
-
-
-
-#TODO login form to be created
-@auth.route('/login', methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data.lower()).first()
-        if user is not None and user.verify_password(form.password.data):
-            login_user(user)
-            next = request.args.get('next')
-            if next is None or not next.startswith('/'):
-                next = url_for('main.index')
-            return redirect(next)
-        flash('Invalid username or password.')
-    return render_template('auth/login.html', form=form)
-
-
-@auth.route('/logout', methods=['GET', 'POST'])
-@login_required
-def logout():
-    logout_user()
-    flash('you are no longer logged in')
-    return redirect(url_for('main.index'))
-
-
-#TODO register form to be created
-@auth.route('/register', methods=['GET', 'POST'])
-def register():
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        pass
-        return redirect(url_for('main.index'))
-    return render_template('auth/register.html', form=form)
-
-#TODO add User profile page use user.html file for tha
