+++ reuse_previous_version/__init__.py
@@ -1,0 +1,63 @@
+from flask import Flask
+from config import Config
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
+from flask_login import LoginManager
+from app import models, routes
+
+app = Flask(__name__)
+app.config.from_object(Config)
+db = SQLAlchemy(app)
+migrate = Migrate(app, db)
+login = LoginManager(app)
+
+
+from .main import main as main_blueprint
+app.register_blueprint(main_blueprint)
+
+from .auth import auth as auth_blueprint
+app.register_blueprint(auth_blueprint, url_prefix='/auth')
+
+
+
+
+
+
+
+
+
+# from flask import Flask
+# from flask_bootstrap import Bootstrap
+# from flask_moment import Moment
+# from flask_sqlalchemy import SQLAlchemy
+# from flask_mail import Mail
+# from flask_login import LoginManager
+# from config import config
+
+
+# bootstrap = Bootstrap()
+# moment = Moment()
+# db = SQLAlchemy()
+# mail = Mail()
+
+# login_manager = LoginManager()
+# login_manager.login_view = 'auth.login'
+
+# def create_app(config_name):
+#     app = Flask(__name__)
+#     app.config.from_object(config[config_name])
+#     config[config_name].init_app(app)
+
+#     bootstrap.init_app(app)
+#     moment.init_app(app)
+#     mail.init_app(app)
+#     db.init_app(app)
+#     login_manager.init_app(app)
+ 
+#     from .main import main as main_blueprint
+#     app.register_blueprint(main_blueprint)
+
+#     from .auth import auth as auth_blueprint
+#     app.register_blueprint(auth_blueprint, url_prefix='/auth')
+
+#     return app
