+++ reuse_previous_version/auth/__init__.py
@@ -2,4 +2,5 @@ from flask import Blueprint
 
 auth = Blueprint('auth', __name__)
 
-from . import views
+from app.auth import routes
+
