+++ reuse_previous_version/config.py
@@ -1,0 +1,54 @@
+import os
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+
+class Config(object):
+    SECRET_KEY = os.environ.get('SECRET_KEY') or 'supersecretkey'
+    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or \
+        'sqlite:///' + os.path.join(basedir, 'app.db')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+
+
+
+
+
+# basedir = os.path.abspath(os.path.dirname(__file__))
+
+# class Config:
+#     SECRET_KEY = os.environ.get('SECRET_KEY') or 'supersecretkey'
+#     MAIL_SERVER = ('MAIL_SERVER', 'smtp.googlemail.com')
+#     MAIL_PORT = int(os.environ.get('MAIL_PORT', '587'))
+#     MAIL_USE_TLS = os.environ.get('MAIL_USE_TLS', 'true').lower() in \
+#         ['true', 'on', '1']
+#     MAIL_USERNAME = os.environ.get('MAIL_USERNAME')
+#     MAIL_PASSWORD = os.environ.get('MAIL_PASSWORD')
+#     FLASKY_MAIL_SUBJECT_PREFIX = '[Flasky]'
+#     FLASKY_MAIL_SENDER = 'Flasky Admin <flasky@example.com>'
+#     FLASKY_ADMIN = os.environ.get('FLASKY_ADMIN')
+#     SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+
+#     @staticmethod
+#     def init_app(app):
+#         pass
+
+
+# class DevelopmentConfig(Config):
+#     FLASK_DEBUG=True
+#     SQLALCHEMY_DATABASE_URI = os.environ.get('DEV_DATABASE_URL') or \
+#         'sqlite:///' + os.path.join(basedir, 'data-dev.sqlite')
+
+
+# class TestingConfig(Config):
+#     FLASK_DEBUG=True
+#     TESTING = True
+#     SQLALCHEMY_DATABASE_URI = os.environ.get('TEST_DATABASE_URL') or \
+#         'sqlite://'
+
+
+# config = {
+#     'development': DevelopmentConfig,
+#     'testing': TestingConfig,
+#     'default': DevelopmentConfig
+# }
