+++ reuse_previous_version/main.py
@@ -1,0 +1,23 @@
+import os
+import click
+from flask_migrate import Migrate, upgrade
+from app import create_app, db
+from app.models import User, Ingredient, Container
+
+app = create_app()
+
+# app = create_app(os.getenv('FLASK_CONFIG') or 'default')
+# migrate = Migrate(app, db)
+
+
+# @app.shell_context_processor
+# def make_shell_context():
+#     return dict(db=db, User=User, Ingredient=Ingredient,
+#                 Container=Container)
+
+# @app.cli.command()
+# def test():
+#     """Run the unit tests."""
+#     import unittest
+#     tests = unittest.TestLoader().discover('tests')
+#     unittest.TextTestRunner(verbosity=2).run(tests)
