+++ reuse_previous_version/main/__init__.py
@@ -4,5 +4,5 @@ from flask import Blueprint
 
 main = Blueprint('main', __name__)
 
-from . import views, errors
+from . import routes, errors
